Split code into submodules
